requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
fastapi==0.119.1
h11==0.16.0
idna==3.11
pillow==12.0.0
pydantic==2.12.3
pydantic_core==2.41.4
qrcode==8.2
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
